Extend check-location to also compare basenames
The previous version of check-location examined the current directory (as
a proxy for the directory in which the file calling check-location
appears) to check if it is of the form */pdp-*/setNN but did not check
the supplied base file name.  This version uses the actual path to the
calling file and additionally checks that the basename of the calling file
matches the file name supplied to check-location.

In order to obtain the path to the calling file, check-location is now a
macro that captures the file's path and expands into a call to the function
check-location-actual.